requirements.txt
flask
jinja2
gunicorn
flask-login
flask-sqlalchemy
psycopg2-binary
